Refactor benchmark package to better separate different kinds of tests
