+++ tools/benchmark/src/MochaReporter.ts
@@ -12,7 +12,7 @@ import type { BenchmarkResult, BenchmarkError } from "./ResultTypes";
 // TODO: this file should be moved in with the mocha specific stuff, but is left where it is for now to avoid breaking users of this reporter.
 // Since it's not moved yet, it needs this lint suppression to do this import:
 // eslint-disable-next-line import-x/no-internal-modules
-import { getName, getSuiteName } from "./mocha/mochaReporterUtilities";
+import { getName, getSuiteName } from "./mocha/mochaReporterUtilities.js";
 
 /**
  * Custom mocha reporter (can be used by passing the JavaScript version of this file to mocha with --reporter).
