+++ tools/benchmark/src/durationBenchmarking/README.md
@@ -1,0 +1,4 @@
+# Duration Benchmarking
+
+Benchmarking which measures how long it takes to run batches of iterations of the benchmark function,
+and automatically adjusts the batch size to try to get accurate results in a reasonable time.
